requirements.txt
# Core
streamlit>=1.32.0
groq>=0.5.0
pydantic>=2.6.0
python-dotenv>=1.0.0

# PDF parsing
pymupdf>=1.24.0         # baseline + figure cropping
marker-pdf>=0.2.0       # high-accuracy PDF→markdown with LaTeX  (pip install marker-pdf)

# Report generation
reportlab>=4.1.0
matplotlib>=3.7.0       # equation rendering (mathtext)

# Gemini vision (no SDK needed — uses urllib only)
# Get free key: https://aistudio.google.com/app/apikey
